# car_rental/swagger.yml
openapi: 3.0.2
info:
  title: Car Rental API
  version: 1.2.0
  description: A simple Car Rental API using Flask-RESTX
servers:
  - url: http://195.148.21.89:5000/api/
  - url: http://127.0.0.1:5000/api/
paths:
...
  /rental/{rental_id}:
    get:
      summary: Get a car rented by rental id
      operationId: get_rented_car_rented_by_id
      parameters:
        - in: path
          name: rental_id
          schema:
            type: integer
          required: true
          description: Numeric ID of the customer
      responses:
        200:
          description: Customer rented cars details
          content:
            application/vnd.mason+json:
              schema:
                $ref: '#/components/schemas/Rental'
        404:
          description: Customer rented cars not found
    put:
      summary: updates a rental
      operationId: update_rental
      parameters:
        - in: path
          name: rental_id
          schema:
            type: integer
          required: true
          description: Numeric ID of the rental
      requestBody:
        content:
          application/vnd.mason+json:
            schema:
              $ref: '#/components/schemas/Rental'
      responses:
        204:
          description: rental updated successfully
        400:
          description: Bad request
  /rental/{car_id}:
    get:
      summary: Get a car rented status by car id
      operationId: get_rented_car_status_by_id
      parameters:
        - in: path
          name: car_id
          schema:
            type: integer
          required: true
          description: Numeric ID of the customer
      responses:
        200:
          description: Customer rented cars details
          content:
            application/vnd.mason+json:
              schema:
                $ref: '#/components/schemas/Rental'
        404:
          description: Customer rented cars not found
...
components:
  schemas:
...
    Rental:
      type: object
      properties:
        id:
          type: integer
          description: Car ID
        car_id:
          type: integer
          description: Car id
        customer_id:
          type: integer
          description: Customer id
        status:
          type: string
          description: Car status
        from_date:
          type: string
          description: starting date
        to_date:
          type: string
          description: end date
        available_from:
          type: string
          description: car available date

      required:
        - id
        - car_id
        - customer_id
        - status
        - from_date
        - to_date
        - available_from
      example:
        id: 0
        car_id: 0
        customer_id: 0
        status: "string"
        from_date: "string"
        to_date: "string"
        available_from: "string"